requirements.txt
fastapi
uvicorn[standard]
pandas
boto3
sqlalchemy
psycopg2-binary
python-dotenv
duckdb